worker/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
soundfile==0.12.1
